requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
pandas==2.1.3
pytz==2023.3
python-multipart==0.0.6 